+++ ts/packages/providers/anthropic/src/index.ts
@@ -14,9 +14,8 @@ import {
   ExecuteToolFnOptions,
   ToolExecuteParams,
   logger,
-  McpProvider,
-  McpServerGetResponse,
   McpUrlResponse,
+  McpServerGetResponse,
 } from '@composio/core';
 import Anthropic from '@anthropic-ai/sdk';
 import { AnthropicTool, InputSchema } from './types';
@@ -49,11 +48,9 @@ export type AnthropicContentBlock = {
  */
 export class AnthropicProvider extends BaseNonAgenticProvider<
   AnthropicToolCollection,
-  AnthropicTool,
-  McpServerGetResponse
+  AnthropicTool
 > {
   readonly name = 'anthropic';
-  readonly mcp: McpProvider<McpServerGetResponse>;
   private chacheTools: boolean = false;
 
   /**
@@ -84,39 +81,7 @@ export class AnthropicProvider extends BaseNonAgenticProvider<
   constructor(options?: { cacheTools?: boolean }) {
     super();
     this.chacheTools = options?.cacheTools ?? false;
-    // Use the base provider's createMcpProvider helper method
-    this.mcp = this.createMcpProvider();
     logger.debug(`AnthropicProvider initialized [cacheTools: ${this.chacheTools}]`);
-  }
-
-  /**
-   * Transform MCP URL response into Anthropic-specific format.
-   * Uses the default transformation from base McpProvider.
-   */
-  transformMcpResponse(
-    data: McpUrlResponse,
-    serverName: string,
-    connectedAccountIds?: string[],
-    userIds?: string[],
-    toolkits?: string[]
-  ): McpServerGetResponse {
-    if (connectedAccountIds?.length && data.connected_account_urls) {
-      return data.connected_account_urls.map((url: string, index: number) => ({
-        url: new URL(url),
-        name: `${serverName}-${connectedAccountIds[index]}`,
-        toolkit: toolkits?.[index],
-      })) as McpServerGetResponse;
-    } else if (userIds?.length && data.user_ids_url) {
-      return data.user_ids_url.map((url: string, index: number) => ({
-        url: new URL(url),
-        name: `${serverName}-${userIds[index]}`,
-        toolkit: toolkits?.[index],
-      })) as McpServerGetResponse;
-    }
-    return {
-      url: new URL(data.mcp_url),
-      name: serverName,
-    } as McpServerGetResponse;
   }
 
   /**
@@ -325,5 +290,44 @@ export class AnthropicProvider extends BaseNonAgenticProvider<
     }
 
     return outputs;
+  }
+
+  /**
+   * Transform MCP URL response into Anthropic-specific format.
+   * By default, Anthropic uses the standard format (same as default),
+   * but this method is here to show providers can customize if needed.
+   *
+   * @param data - The MCP URL response data
+   * @param serverName - Name of the MCP server
+   * @param connectedAccountIds - Optional array of connected account IDs
+   * @param userIds - Optional array of user IDs
+   * @param toolkits - Optional array of toolkit names
+   * @returns Standard MCP server response format
+   */
+  transformMcpResponse(
+    data: McpUrlResponse,
+    serverName: string,
+    connectedAccountIds?: string[],
+    userIds?: string[],
+    toolkits?: string[]
+  ): McpServerGetResponse {
+    // Anthropic uses the standard format
+    if (connectedAccountIds?.length && data.connected_account_urls) {
+      return data.connected_account_urls.map((url: string, index: number) => ({
+        url: new URL(url),
+        name: `${serverName}-${connectedAccountIds[index]}`,
+        toolkit: toolkits?.[index],
+      })) as McpServerGetResponse;
+    } else if (userIds?.length && data.user_ids_url) {
+      return data.user_ids_url.map((url: string, index: number) => ({
+        url: new URL(url),
+        name: `${serverName}-${userIds[index]}`,
+        toolkit: toolkits?.[index],
+      })) as McpServerGetResponse;
+    }
+    return {
+      url: new URL(data.mcp_url),
+      name: serverName,
+    } as McpServerGetResponse;
   }
 }
